--- a/paddle/fluid/pybind/parallel_executor.cc
+++ b/paddle/fluid/pybind/parallel_executor.cc
@@ -689,6 +689,34 @@ void BindParallelExecutor(pybind11::module &m) {  // NOLINT
                         build_strategy.sequential_run = True
                      )DOC")
       .def_property(
+          "fuse_dot_product_attention",
+          [](const BuildStrategy &self) {
+            return self.fuse_dot_product_attention_;
+          },
+          [](BuildStrategy &self, bool b) {
+            PADDLE_ENFORCE_NE(self.IsFinalized(),
+                              true,
+                              platform::errors::PreconditionNotMet(
+                                  "BuildStrategy has been finlaized, cannot be "
+                                  "configured again."));
+            self.fuse_dot_product_attention_ = b;
+          },
+          R"DOC((bool, optional): fuse_dot_product_attention indicate whether
+                to fuse dot product attention,
+                it would make the execution faster. Default is False.
+
+                Examples:
+                    .. code-block:: python
+
+                        import paddle
+                        import paddle.static as static
+
+                        paddle.enable_static()
+
+                        build_strategy = static.BuildStrategy()
+                        build_strategy.fuse_dot_product_attention = True
+                     )DOC")
+      .def_property(
           "fuse_bn_act_ops",
           [](const BuildStrategy &self) { return self.fuse_bn_act_ops_; },
           [](BuildStrategy &self, bool b) {