--- a/paddle/phi/kernels/gpudnn/mha_cudnn_frontend.h
+++ b/paddle/phi/kernels/gpudnn/mha_cudnn_frontend.h
@@ -1,0 +1,84 @@
+// Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#ifdef PADDLE_WITH_CUDNN_FRONTEND
+#include "paddle/phi/backends/dynload/cudnn_frontend.h"
+
+enum class MHA_Layout {
+  NOT_INTERLEAVED = 0,
+  QKV_INTERLEAVED = 1,
+  KV_INTERLEAVED = 2
+};
+
+enum class MHA_Matrix {
+  Q_Matrix = 0,  // queries
+  K_Matrix = 1,  // keys
+  V_Matrix = 2,  // values
+  S_Matrix = 3,  // output of GEMM1
+  O_Matrix = 4,  // final output
+};
+
+enum class MHA_Bias_Type {
+  NO_BIAS = 0,
+  PRE_SCALE_BIAS = 1,
+  POST_SCALE_BIAS = 2
+};
+
+void run_cudnn_fmha_fwd(int64_t b,
+                        int64_t h,
+                        int64_t s_q,
+                        int64_t s_kv,
+                        int64_t d,
+                        int64_t seed,
+                        MHA_Layout layout,
+                        float scaling_factor,
+                        double dropout_probability,
+                        MHA_Bias_Type bias_type,
+                        bool is_causal_masking,
+                        void *devPtrQ,
+                        void *devPtrK,
+                        void *devPtrV,
+                        void *devPtrS,
+                        void *devPtrO,
+                        void *devPtrBias,
+                        void *devActualSeqlenQ,
+                        void *devActualSeqlenKV,
+                        cudnnDataType_t tensorType,
+                        cudnnHandle_t handle_);
+
+void run_cudnn_fmha_bwd(int64_t b,
+                        int64_t h,
+                        int64_t s_q,
+                        int64_t s_kv,
+                        int64_t d,
+                        MHA_Layout layout,
+                        float scaling_factor,
+                        float dropout_probability,
+                        bool is_causal_masking,
+                        void *devPtrQ,
+                        void *devPtrK,
+                        void *devPtrV,
+                        void *devPtrS,
+                        void *devPtrdQ,
+                        void *devPtrdK,
+                        void *devPtrdV,
+                        void *devPtrdO,
+                        void *devPtrdS,
+                        void *devActualSeqlenQ,
+                        void *devActualSeqlenKV,
+                        cudnnDataType_t tensorType,
+                        cudnnHandle_t handle_);
+#endif