… paddle/fluid/framework/ir/fuse_dot_product_attention_pass.cc
… paddle/fluid/framework/ir/graph_pattern_detector.cc
… paddle/fluid/operators/fused/fused_dot_product_attention_op.cc
… paddle/fluid/operators/fused/fused_dot_product_attention_op.cu
… paddle/fluid/pybind/eager_generator.h
… paddle/phi/kernels/gpudnn/mha_cudnn_frontend.cu
… patches/cudnn-frontend/0001-patch-for-paddle.patch
… python/paddle/nn/layer/transformer.py
… test/legacy_test/test_fuse_dot_product_attention_pass.py
… test/legacy_test/test_fused_dot_product_attention_op.py