<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
New features

### PR changes
OPs

### Description
CUDNN v8 provide fused MHA kernel to accelerate MultiHeadAttention(refer to [cudnn-frontend/fused_mha_sample](https://github.com/NVIDIA/cudnn-frontend/blob/main/samples/fused_mha_sample.cpp)). To be exact, it only fuse the calculation of _**scaled dot product attention**_, which does not include **_qkv projection_** and **_output projection_**, but only includes **_scale, bmm1, mask, softmax, dropout, and bmm2_**. So, let's call it **_fused_dot_product_attention_** instead of **_fused_multi_head_attention_**.

1. Added the following OPs to compute **_scaled dot product attention_**  (forward + backward):
    - fused_dot_product_self_attention/fused_dot_product_self_attention_grad
    - fused_dot_product_cross_attention/fused_dot_product_cross_attention_grad
2. Added 2 optional arguments in paddle.nn.MultiHeadAttention: **_fuse_qkv_** and **_mode_**
    - Default: **_fuse_qkv_**==False and **_mode_** == None,  the QKV projection will use 3 different weight tensor.
    - If **_fuse_qkv_**==True and **_mode_** == "self_attn", the QKV projection will use a single weight tensor with shape [embed_dim, 3*embed_dim], and the output of QKV projection can be reshaped into [b, s, 3, h, d].
    - If **_fuse_qkv_**==True and **_mode_** == "cross_attn", the KV projection will use a single weight tensor with shape [embed_dim, 2*embed_dim], and the output of KV projection can be reshaped into [b, s, 2, h, d].
3. Added a new IR pass to detect above "self_attn" pattern and "cross_attn" pattern. 
4. Added a class member fuse_dot_product_attention to BuildStrategy to enable fuse_dot_product_attention_pass.

**Note:** 
The difference between the two attention mode is that the input of self-attention requires QKV to be a single buffer, while the input of cross-attention only requires KV to be a single buffer, with Q existing separately in another buffer.   
More specific details are that the **_QKV_** input for **_fused_dot_product_self_attention_** op has shape **_[b, s, 3, h, d]_**, where _**[:, :, 0, :, :]**_ for **_Q_**,  **_[:, :, 1, :, :]_** for **_K_**,  **_[:, :, 2, :, :]_** for **_V_**. The KV input for **_fused_dot_product_cross_attention_** op has shape **_[b, s, 2, h, d]_**, where **_[:, :, 0, :, :]_** for **_K_**,  **_[:, :, 1, :, :]_** for **_V_**.

**Usage sample (CUDNN >= 8.9.1 is required):**
```
class TransformerLayer(paddle.nn.Layer):
    def __init__(self, hidden, num_heads):
        super().__init__()
        self.encoder_layer = paddle.nn.TransformerEncoderLayer(hidden, num_heads, 4*hidden, dropout=0.1, fuse_qkv=True)
        self.encoder = paddle.nn.TransformerEncoder(self.encoder_layer, 2)

    def forward(self, q, mask):
        out = self.encoder(q, mask)
        loss = paddle.mean(out)
        return loss

paddle.enable_static()
main_prog = paddle.static.Program()
startup_prog = paddle.static.Program()

batch_size = 32
head_size = 64
num_head = 12
hidden = head_size * num_heads
seq_len = 128

with paddle.static.program_guard(main_prog, startup_prog):
    enc_input = paddle.static.data(name='enc_input', shape=[-1, -1, hidden], dtype="float32")
    attn_mask = paddle.static.data(name='mask', shape=[-1, 1, 1, seq_len], dtype="int32")
    model = TransformerLayer(hidden, num_heads)
    loss = model(enc_input, attn_mask)
    opt = paddle.optimizer.SGD(learning_rate=0.1)
    amp_list = paddle.static.amp.CustomOpLists(custom_white_list=['softmax'])
    opt = paddle.static.amp.decorate(
        optimizer=opt,
        amp_lists=amp_list,
        init_loss_scaling=128.0,
        use_dynamic_loss_scaling=True)
    opt.minimize(loss)


place = paddle.CUDAPlace(0)
exe = paddle.static.Executor(place)
exe.run(startup_prog)

build_strategy = paddle.static.BuildStrategy()
build_strategy.fuse_dot_product_attention = True
main_prog = paddle.static.CompiledProgram(main_prog)
main_prog = main_prog.with_data_parallel(loss_name=loss.name,
                                         build_strategy=build_strategy,
                                         places=place)
exe.run(main_prog, feed=feed)
# 2 subgraphs be fused to fused_dot_product_self_attention op
# 2 subgraphs be fused to fused_dot_product_self_attention_grad op
```


BERT Perf test (8 x A100 80GB):
| seqlen | original (sequence per second) |  cuDNN MHA (sequence per second) | speedup|
| ------- | ------- | ------- | -------|
| 128     |   610 |   685 |  1.12x |
| 512     |   116 |   157 |  1.35x |

BERT convergence test (8 x A100 80GB x 32 nodes):
- Phase1 final loss: 1.34
- Phase2 final loss: 1.24
- SQuAD F1: 90.5

BERT code:
[https://github.com/NVIDIA/DeepLearningExamples/tree/master/PaddlePaddle/LanguageModeling/BERT](https://github.com/NVIDIA/DeepLearningExamples/tree/master/PaddlePaddle/LanguageModeling/BERT)