@JamesLim-sy 2023-06-07
> # IR pass
> 
> IR pass部分的实现较为复杂，这里做一些解释以便review
> ## 设计思路:
> 
> **QKV输入**：fused_dot_product_attention 不计算qkv projection和out projection，只计算中间的scale, bmm1, mask, softmax, dropout, bmm2等。并且要求输入的qkv(self attention)或者kv(cross attention)是同一块buffer。但nn.MultiHeadAttention中的qkv projection的输出是三个独立的q, k, v tensor，不能直接作为fused_dot_product_attention的输入。因此我们对nn.MultiHeadAttention新增了两个optional的参数：fuse_qkv和mode。
> 
>     1. 若fuse_qkv=False, mode=None, 将按照paddle默认的实现，将qkv分开计算
> 
>     2. 若fuse_qkv=True, mode=”self_attn”, qkv projection的weight合成一整块，shape为[hidden_dim, 3_hidden_dim]来计算，输出shape为[b, s, 3_hidden_dim]，reshape成[b, s, 3, h, d]，作为fused_dot_product_self_attention的输入。
> 
>     3. 若fuse_qkv=True, mode=”cross_atten”, kv projection的weight合成一整块，q projection单独计算。则输出reshape成两个：q→[b, s, h, d]  kv→[b, s, 2, h, d], 作为fused_dot_product_cross_attention的输入。
> 
> 
> 由于我们的设计是用IR pass来对subgraph进行遍历和替换，并没有直接在nn.MultiHeadAttention的forward中调用fused op，所以融合的qkv或kv会被再次slice成独立的tensor，然后来按paddle native mha的流程来计算。如下左图。
> 
> 仅当用户开启build_strategy.fuse_dot_product_attention=True（并且系统满足cudnn版本等要求）之后，这些slice op连同后面的dot_product_attention的subgraph会被替换成fused op。如下右图。
> 
> ![image](https://user-images.githubusercontent.com/21985950/243325421-5f17ee10-c279-4ad2-a68b-682ce610d49f.png)
> 
> 注意: 由于fused_dot_product_attention op并不能直接输入mask，而是输入actual_seqlen，所以需要插入额外的squeeze, reduce_sum等op将mask转化为actual_seqlen。
> ## 实现细节：
> 
> **fuse_dot_product_attention_pass.h:** **fuse_dot_product_attention_pass.cc**
> 
>     1. **3个OpCache**
>        
>        1. SequenceMetaCache: 用来存actual_seqlen节点（等同于mask），比如BERT large有24个layer，对同一个输入batch来说，计算24次mha的fwd/bwd都会使用同一组mask，所以只需要计算一次actual seqlen然后存在cache中。
>        2. SoftmaxOutputCache：cudnn mha中bwd的计算需要fwd的中间结果，即softmax的输出（若有dropout，则为dropout的输出）。所以fwd计算完softmax，存储在cache中，然后bwd再取出来用。
>        3. QKVCache：同理，fwd计算时将qkv存储在cache中，在对应的bwd计算中取出来，用来计算dq, dk, dv。
> 
>     2. **nodes_to_remove**
>        如上所述，以BERT-large为例，需要将24个fwd/bwd中所有的subgraph都替换成fused op。由于某些node的指针存储在cache中，在bwd时要再次使用，所以不能每次替换一个subgraph后就立即删除掉被替换的节点，而是替换完所有的subgraph后再删除。
> 
>     3. **InsertActualSeqlenOp_**
>        插入一些小的paddle op将mask转化成q_actual seqlen和k_actual_seqlen，目前支持两种情况：
>        
>        1. mask shape: [b, 1, 1, s], BERT encoder中常见的shape，需要转化成q_actual_seqlen和k_actual_seqlen，首先插入**squeeze**，去掉中间的两个1维度变成[b, s]，然后用ones_like拷贝一个全为1的[b, s] tensor。对这两个tensor分别做reduce_sum就可以得到shape为[b, ]的q_actual_seqlen和k_actual_seqlen。
>           这种情况下插入了squeeze, ones_like, reduce_sum, reduce_sum 4个op。实际上这种mask等效于只在k上做了mask，所以q_actual_seqlen的值都是max_seqlen
>        2. mask shape: [b, 1, s_q, s_kv], decoder中常用的shape。类似的，等效于q和k都需要mask，需要插入slice, slice, reduce_sum, reduce_sum 4个op。
>        3. 未来若有需要，应当支持causal masking，暂未实现

请问下。这块存储的3个OpCache，是否与`flash_attention`中的`softmax_lse`, `fmha_out`, `seq_len_r` 的三种前向计算的驻留信息相仿。