@Wong4j 2023-06-08
> 请问下。这块存储的3个OpCache，是否与`flash_attention`中的`softmax_lse`, `fmha_out`, `seq_len_r` 的三种前向计算的驻留信息相仿。

我目前还不熟悉flash attention的细节。参考cudnn frontend 的 sample： [fused_mha_sample.cpp](https://github.com/NVIDIA/cudnn-frontend/blob/v0.8/samples/fused_mha_sample.cpp)

- actual_seqlen很好理解，比如
```
# first batch.
shape: [2, 8]
mask: [[1, 1, 1, 1, 1, 1, 1, 0],
       [1, 1, 1, 1, 1, 0, 0, 0]]
# second batch
shape: [2, 16]
mask: [[1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0],
       [1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0]]
```
第一个batch的 max_seqlen = 8, actual seqlen = [7, 5]
第二个batch的 max_seqlen = 16, actual seqlen = [13, 8]

- SoftmaxOutputCache
forward计算出来的中间结果`S`，在backward需要作为输入，如下图
![image](https://github.com/PaddlePaddle/Paddle/assets/21985950/0e8cd685-5f60-4a79-8213-31a594671d16)

- QKVCache
如上图，forward需要Q, K, V作为输入，backward也需要Q, K, V作为输入，用来计算dQ, dK, dV。所以在计算forward的时候把QKV存到cache中，backward可以直接取出来用。