@Wong4j 2023-11-06
I have submit a new PR https://github.com/PaddlePaddle/Paddle/pull/58680. Close this PR